Refactor runtests to allow TestEnv
